Add !ask command
This command is to help cut down on the amount of times we need to type "You don't need to ask to ask. Just ask your question" in the discord chat.

Note: See comment on line 74. I'm sure that needs to be moved elsewhere, I'm just not seeing where it should go.